+++ src/ai-bundle/config/options.php
@@ -53,6 +53,7 @@ return static function (DefinitionConfigurator $configurator): void {
                     ->append($import('platform/openai'))
                     ->append($import('platform/openresponses'))
                     ->append($import('platform/openrouter'))
+                    ->append($import('platform/orqai'))
                     ->append($import('platform/ovh'))
                     ->append($import('platform/perplexity'))
                     ->append($import('platform/scaleway'))